Moved helpers for newsletter emails into a separate util
no issue
